Extract AI workflow demo into reusable component and remove custom animations
- Move AIWorkflows demo logic from lifetime-deal page into new AiWorkflowDemo component
- Relocate AiMessage, WorkflowCanvas, and InputArea to components/AiWorkflowDemo directory
- Add compact prop to AiWorkflowDemo for flexible sizing
- Replace custom CSS animations (aiwf-glow-pulse, aiwf-dot-pulse, aiwf-blink, aiwf-float) with Tailwind's animate-pulse
- Remove glow animation from background gradient
- Change intersection || 1
